pyproject.toml
[tool.poetry]
name = "hugr-llvm"
version = "0.1.0"
description = ""
authors = ["Douglas Wilson"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
guppylang = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
